backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
firebase-admin==6.5.0
pydantic==2.10.0
pydantic-settings==2.6.0
python-dotenv==1.0.1
pytest==8.3.0
pytest-asyncio==0.24.0
hypothesis==6.122.0
httpx==0.28.0
